Refactor: improve type handling, simplify assertions, and apply minor formatting adjustments across several modules.

+++ ualbf-project/rust-engine/src/dfs_tree.rs
@@ -169,7 +169,7 @@ fn explore_prefix(
     // Run this only if we have at least 3 primes to form a meaningful lattice geometry
     if curr.factors.len() >= 3 {
         // We pass curr.n_l to give the lattice the widest, safest tolerance
-        if crate::lattice::lll_prune_prefix(&curr.factors, curr.n_l as u128) {
+        if crate::lattice::lll_prune_prefix(&curr.factors, curr.n_l) {
             abundance_pruned.fetch_add(1, Ordering::Relaxed);
             return;
         }
